refactor: dedupe report-store helpers and trend math; drop dead expr
#95: findJsonFiles/loadReport/isGraded (plus StoredReport/GradeDistribution)
were byte-identical in export.ts and aggregate.ts. Extract them to a new
src/report-store.ts imported by both; aggregate.ts re-exports isGraded so
existing consumers/tests keep working. Remove the dead
`(data.allUngraded ? 0 : 0)` no-op in buildLanguages and the vestigial
allUngraded accumulator, computing language count once from totalCounts.

#94: trends.astro reimplemented the prev-vs-current delta math already in
computeTrendSummary. Factor the pure trend math into a side-effect-free
src/trend-summary.ts (computeRepoDeltas + computeTrendSummary); trend-view.ts
re-exports it for backward compatibility, and the Astro page imports it at
build time and maps TrendDelta onto its view model. Rendered /trends output
verified byte-identical (only the build-nondeterministic CSS hash differs).

Fixes #94
Fixes #95
